refactor: clean up type definition for parser fn
