+++ src/builders/types/index.ts
@@ -1,4 +1,3 @@
 export * from './date-parser.fn';
 export * from './date-time.options';
-export * from './parser.fn';
 export * from './time.input';
