--- src/builders/types/parser.fn.ts
@@ -1,3 +1,0 @@
-import { Nullable } from '../../types';
-
-export type ParserFn<T, K, R> = (input: T, opts: K) => Nullable<R>;
