+++ src/parser/types/parser.fn.ts
@@ -1,0 +1,3 @@
+import { Nullable } from '../../types';
+
+export type ParserFn<T, K, R> = (input: T, opts?: K) => Nullable<R>;
