+++ src/types/index.ts
@@ -1,2 +1,1 @@
 export * from './nullable';
-export * from './parser.fn';
