--- src/types/parser.fn.ts
@@ -1,1 +1,0 @@
-export type ParserFn<T, K = T> = (input: T, ...params: any[]) => K;
